+++ inst/paper/paper.bib
@@ -13,11 +13,12 @@
     year = {2021},
     note = {R package version 2.5.1},
     url = {https://CRAN.R-project.org/package=R6},
+    doi = {10.32614/CRAN.package.R6},
 }
 
 @misc{
     OpenMpp,
-    author = {Matthew T. Warkentin},
+    author = {{OpenM++ Contributors}},
     title = {{OpenM++: open source microsimulation platform}},
     year = {2024},
     howpublished = {https://openmpp.org/index.html},
