+++ inst/paper/paper.md
@@ -26,12 +26,18 @@ bibliography: paper.bib
 
 # Summary
 
+OpenM++ is a powerful open-source and cross-platform software and domain-specific language to develop, use, and deploy dynamic microsimulations or agent-based models. The [`openmpp`](https://github.com/mattwarkentin/openmpp) R package provides a programmatic interface to OpenM++ to simplify the process of connecting to local or remote servers with secure user authentication, creating, managing, and running large-scale and complex microsimulations, performing scalable and reproducible computations, and gathering results for reproducible reporting.
+
+# Statement of Need
+
 Dynamic microsimulation is a computational technique used to simulate the behavior of individual entities (i.e., micro) within a population over time. It combines elements of microeconomics, demography, and statistical modeling to create detailed simulations of complex systems. Dynamic microsimulation involves creating a synthetic population representative of the real population and modeling the life paths of individuals based on various demographic and socioeconomic characteristics. Unlike static models that provide snapshots of a population at a single point in time, dynamic microsimulation models capture the interactions and changes that occur over time, allowing researchers to analyze the long-term effects of demographic shifts, policy changes, as well as clinical or population intervention. These models can then be used to predict future population trends and assess the impact of policy changes or interventions on health, resource, and economic outcomes. Dynamic microsimulation is used across a wide range of disciplines, including economics, public policy, epidemiology, urban planning, and environmental science. In public health research, dynamic microsimulation models are used to forecast disease trends, evaluate the impact of healthcare interventions, and plan for future healthcare needs. Overall, dynamic microsimulation provides a powerful framework for understanding the complex dynamics of trends within a population to inform evidence-based decision-making across a variety of fields.
 
 [OpenM++](https://openmpp.org/index.html) is an open-source software platform designed for dynamic microsimulation modeling [@OpenMpp]. It provides a flexible framework and language for researchers and policymakers to build, calibrate, and analyze complex dynamic microsimulation models. OpenM++ offers a user-friendly interface and a powerful domain-specific language, allowing users to define intricate demographic, economic, and behavioral processes within a simulated population. With its modular architecture, OpenM++ enables easy integration of various data sources and model components, facilitating interdisciplinary research and policy analysis. The OpenM++ platform is widely used in academia, government agencies, and research institutions to explore the long-term impacts of policy interventions, demographic changes, and economic trends on populations and societies. However, while the OpenM++ graphical interface is simple to learn, the graphical interface may impede users with more complicated modelling needs and limits the transparency and reproducibility of analyses. The OpenM++ software provides an application programming interface (API) that enables programmatic interactions with data and models using scripting languages (e.g., R).
 
[…]
