+++ pkgdown.yml
@@ -2,7 +2,7 @@ pandoc: 3.1.11
 pkgdown: 2.1.0
 pkgdown_sha: ~
 articles: {}
-last_built: 2024-08-26T20:26Z
+last_built: 2024-08-26T22:00Z
 urls:
   reference: https://mattwarkentin.github.io/openmpp/reference
   article: https://mattwarkentin.github.io/openmpp/articles
