### 2024-10-03 #7146 @mattwarkentin
Hi @mikemahoney218! Glad to see you as the editor for this submission. Most of the people I am familiar with who are expert in this field are not really active GitHub users as far as I can tell.

From the JOSS reviewer database, the following people look potentially suitable (sometimes hard to tell):
1. [agerada](https://github.com/agerada)
2. [yafeiwang89](https://github.com/yafeiwang89)
3. [jkbren](https://github.com/jkbren)
4. [pausz](https://github.com/pausz)
5. [tpoisot](https://github.com/tpoisot)
