### 2024-10-17 #7146 @mattwarkentin
I think someone with expertise in OpenM++ will be harder to find and maybe not all that necessary. I think general microsimulation knowledge and expertise in R package development and communicating with APIs may be more than enough.
