### 2024-11-07 #7435 @editorialbot
[REVIEW]: openmpp: An R package for dynamic microsimulation modeling using OpenM++
**Submitting author:** @mattwarkentin (<a href="http://orcid.org/0000-0001-8730-3511">Matthew T. Warkentin</a>)
**Repository:** https://github.com/mattwarkentin/openmpp
**Branch with paper.md** (empty if default branch):
**Version:** 0.0.1
