### 2024-11-12 fc1d84c
Fixes issue where requiring base digest was too strict and other QoL fixes
