### 2024-11-12 1184617
Better admin controls to clean up memory
