+++ man/admin_models_refresh.Rd
@@ -1,0 +1,53 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/Admin.R
+\name{admin_models_refresh}
+\alias{admin_models_refresh}
+\alias{admin_models_close}
+\alias{admin_database_close}
+\alias{admin_database_open}
+\alias{admin_database_cleanup}
+\alias{admin_cleanup_logs}
+\alias{admin_cleanup_log}
+\alias{admin_models_pause}
+\alias{admin_models_pause_all}
+\alias{admin_service_shutdown}
+\title{OpenM++ Administrative Tasks}
+\usage{
+admin_models_refresh()
+
+admin_models_close()
+
+admin_database_close(model)
+
+admin_database_open(path)
+
+admin_database_cleanup(path, name = NULL, digest = NULL)
+
+admin_cleanup_logs()
+
+admin_cleanup_log(name)
+
+admin_models_pause(pause)
+
+admin_models_pause_all(pause)
+
+admin_service_shutdown()
+}
+\arguments{
+\item{model}{Model digest or model name.}
+
+\item{path}{Path to model database file relative to the \code{models/bin} folder.
+For example, the name of the model with a \code{".sqlite"} extension.}
+
+\item{name}{Model name. Optional for \code{admin_cleanup_db()}.}
+
+\item{digest}{Model digest. Optional for \code{admin_cleanup_db()}.}
+
+\item{pause}{Logical. Whether to pause or resume model runs queue processing.}
+}
+\value{
+A \code{list} or nothing, invisibly.
+}
+\description{
+Functions for performing administrative tasks.
+}
