+++ man/admin_refresh_models.Rd
@@ -1,29 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/Admin.R
-\name{admin_refresh_models}
-\alias{admin_refresh_models}
-\alias{admin_close_models}
-\alias{admin_pause_models}
-\alias{admin_pause_all_models}
-\alias{admin_shutdown_service}
-\title{OpenM++ Administrative Tasks}
-\usage{
-admin_refresh_models()
-
-admin_close_models()
-
-admin_pause_models(pause)
-
-admin_pause_all_models(pause)
-
-admin_shutdown_service()
-}
-\arguments{
-\item{pause}{Logical. Whether to pause or resume model runs queue processing.}
-}
-\value{
-Nothing, invisibly.
-}
-\description{
-Functions for performing administrative tasks.
-}
