+++ NAMESPACE
@@ -5,11 +5,16 @@ export(OpenMppModel)
 export(OpenMppModelRun)
 export(OpenMppModelRunSet)
 export(OpenMppWorkset)
-export(admin_close_models)
-export(admin_pause_all_models)
-export(admin_pause_models)
-export(admin_refresh_models)
-export(admin_shutdown_service)
+export(admin_cleanup_log)
+export(admin_cleanup_logs)
+export(admin_database_cleanup)
+export(admin_database_close)
+export(admin_database_open)
+export(admin_models_close)
+export(admin_models_pause)
+export(admin_models_pause_all)
+export(admin_models_refresh)
+export(admin_service_shutdown)
 export(copy_param_run_to_workset)
 export(copy_param_workset_to_workset)
 export(create_scenario)
