+++ pkgdown.yml
@@ -1,8 +1,8 @@
 pandoc: 3.1.11
-pkgdown: 2.1.0
+pkgdown: 2.1.1
 pkgdown_sha: ~
 articles: {}
-last_built: 2024-08-26T22:00Z
+last_built: 2024-11-12T21:29Z
 urls:
   reference: https://mattwarkentin.github.io/openmpp/reference
   article: https://mattwarkentin.github.io/openmpp/articles
