### 2024-11-12 8add374
Better names for some admin funcs
