+++ NAMESPACE
@@ -10,9 +10,9 @@ export(admin_cleanup_logs)
 export(admin_database_cleanup)
 export(admin_database_close)
 export(admin_database_open)
+export(admin_jobs_pause)
+export(admin_jobs_pause_all)
 export(admin_models_close)
-export(admin_models_pause)
-export(admin_models_pause_all)
 export(admin_models_refresh)
 export(admin_service_shutdown)
 export(copy_param_run_to_workset)
