### 2024-11-19 1e62533
Update docs for how to find path arg for database files
