+++ pkgdown.yml
@@ -2,7 +2,7 @@ pandoc: 3.1.11
 pkgdown: 2.1.1
 pkgdown_sha: ~
 articles: {}
-last_built: 2024-11-12T23:41Z
+last_built: 2024-11-19T02:01Z
 urls:
   reference: https://mattwarkentin.github.io/openmpp/reference
   article: https://mattwarkentin.github.io/openmpp/articles
