### 2024-11-28 mattwarkentin/openmpp#9 @agerada
JOSS software review
Hi @mattwarkentin. I will be adding my review comments on this thread.

A few things to note first. I do not have personal experience with `OpenM++`, so I will predominantly be reviewing this software as an R package. Where possible, I will try to review compatibility with `OpenM++` despite this limitation.

I will probably keep this thread for software comments, and have a separate thread for comments on the JOSS paper itself.

Firstly, `openmpp` appears to be a well-written piece of software, and it was fairly easy for me to interact with it. It also appears to be endorsed (and complemented) by the `OpenM++` team [here](https://openmpp.org/).

Specific comments follow.

1. Nomenclature – there appears to be another R package called `openMpp`: https://github.com/openmpp/R?tab=readme-ov-file. I am not sure whether this is in active development, however this can be confusing to users as names are very similar. @mattwarkentin are you concerned about this? This may not be an issue if there is a plan to retire the old package.
2. Installation and getting started – for users who already have `OpenM++` set up and installed locally, this will not be an an issue. However, for users who are starting off with using `OpenM++` through this R package, the installation instructions are insufficient. Although I did manage to get everything running and installed, I had to search documentation elsewhere (and even then it wasn't that clear on the `OpenM++` site. I would highly recommend a more detailed step-by-step that covers local installation and running of the API (in particular, link to binaries, and how to execute `bin/oms` from shell).
3. Use of R6 as an OOP system. While R6 has advantages for an API-type package like this (which I suspect drove this design choice), unfortunately it does lead to code that is not idiomatic of R code. This package is affected by this limitation – domain users who are approaching this from R (and who do not have experience with other more traditional OOP langauges) will find code such as the following quite difficult to reason about:

```
my_scenario$copy_params('AgeBaselinePreg1')
```

I think that this limitation will have to be accepted, however I would at least suggest expanding a bit more on this in the [readme](https://github.com/mattwarkentin/openmpp/?tab=readme-ov-file#models-scenarios-runs-and-runsets). I would also explain things like 'methods' and 'fields' with alternatives like 'functions' and 'data'. May not be technically correct, but more familiar to R users.
4. Testing suite. Testing the package 'as is' on a local clone fails, which I think is because it expects an active remote API. The testing suite itself does not test any logic, and is mainly asserting the return types of API objects. This is probably acceptable for an API software that is mainly providing an interface. The repository's github actions does appear to be succesfully testing through an API. Could @mattwarkentin clarify, is that using a web-hosted `OpenM++` instance or a temporary one?
5. Developer notes. I cannot find a set of developer notes on how to set up a local developer instance and test the package. That would be useful for users intending to contribute.

Minor points:
1. The link to github [here](https://github.com/mattwarkentin/openmpp/?tab=readme-ov-file#models-scenarios-runs-and-runsets) links to github's main page rather than the repo itself.

I will pause there for now continue the review of the software and paper at a later date.
