### 2024-12-09 mattwarkentin/openmpp#10 @agerada
JOSS paper review
Hi @mattwarkentin.

Here is my review of the JOSS paper submission for `openmpp`.

Any references to line numbers refer to the PDF.

Overall this is a well-written paper and in my opinion fits with JOSS's required paper format.

Some relatively minor comments are:

1) The final section on R6 OOP (lines 61-69) feels a little out of place. I'm not sure it adds much to the paper, but if you do keep it in, I suggest adding a few lines of example code (e.g., load model, load scenario, and execute run).
2) I think more could be added to place this software in the context of related work. Although the paper does say that users can use `OpenM++`'s GUI as an alternative, I would suggest mentioning other alternative scripting approaches to programmatically control modelling, e.g., other R or python packages that interact with `OpenM++`.
