### 2024-12-09 mattwarkentin/openmpp#9 @agerada
Hi @mattwarkentin.

Can you please confirm whether there is a guideline for contributions to this software as per JOSS's requirements:

```
 Community guidelines: Are there clear guidelines for third parties wishing to 1. Contribute to the software 2. Report issues or problems with the software 3. Seek support```

I've had a look through the documentation and didn't find one, but I may have missed it..
