### 2024-12-09 mattwarkentin/openmpp#11 @evaenns
Installation instructions
If the intended audience of this package includes users who are new to OpenM++, then the [installation instructions](https://github.com/mattwarkentin/openmpp?tab=readme-ov-file#user-authentication) could be improved. In particular, more guidance directly to download and start up instructions for OpenM++ would be helpful, more direct access to the `.Renviron` file, and examples of where to find the local API address. This took me a bit of time to figure out. This statement in particular

"If you aren’t sure of your host address, you may contact the OpenM++ administrator to retrieve this information"

wasn't helpful.

Part of: openjournals/joss-reviews#7435
