### 2024-12-09 mattwarkentin/openmpp#12 @evaenns
Example
In the [Example](https://github.com/mattwarkentin/openmpp?tab=readme-ov-file#example) to illustrate `openmpp`, it would be helpful to first have a description of the application to help the user understand what kind of model inputs, model outputs, scenarios, etc. the user should expect to see in the demonstration. Linking to the OpenM++ documentation of the example would be good for additional detail, but a basic description still seems appropriate in the `openmpp` example section.

Part of: https://github.com/openjournals/joss-reviews/issues/7435
