### 2024-12-09 mattwarkentin/openmpp#13 @evaenns
Overall -- specifying audience
If the goal of the package is to only provide additional functionality to existing OpenM++ users, then this should be specified clearly at the top of the package description and made clear in the JOSS paper text. If intended audience also includes people who wish to learn how to use OpenM++ but starting from within R (e.g. `openmpp` as a gateway to using the OpenM++ platform), then more OpenM++ explanation and guidance should be provided.

Part of: https://github.com/openjournals/joss-reviews/issues/7435
