### 2024-12-09 mattwarkentin/openmpp#14 @evaenns
Documentation
Documentation of `openmpp` functions could be improved. Currently it is quite terse. Perhaps it is relying on an understand of OpenM++ nomenclature, but to function as a standalone R package, more plain text description of functions (including as they related to OpenM++ functionality is needed). Many documentation files also include a long list of usage examples, but do not describe what these different uses do. e.g.
https://github.com/mattwarkentin/openmpp/blob/1e62533966619e9641658096395bf607f2d2e050/man/get_workset.Rd#L25C1-L62C40

Part of: https://github.com/openjournals/joss-reviews/issues/7435
