### 2024-12-09 mattwarkentin/openmpp#15 @evaenns
No community guidelines
I did not see any community guidelines as required by JOSS review criteria for third parties wishing to
1. Contribute to the software
2. Report issues or problems with the software
3. Seek support

I did see a code of conduct document.

Part of: https://github.com/openjournals/joss-reviews/issues/7435
