### 2024-12-09 mattwarkentin/openmpp#16 @evaenns
Software paper missing "State of the Field"
The authors do not describe how this software compares to other commonly-used packages. If there are no comparable packages, this should be stated.

Part of: https://github.com/openjournals/joss-reviews/issues/7435
