### 2024-12-09 mattwarkentin/openmpp#17 @evaenns
Difficult to assess functionality without OpenM++ familiarity
The example code provided in the `readme` runs without error and producing expected output. However, it's not clear that this example provides an exhaustive demonstration of the package functionality. Without knowing more about OpenM++, it is difficult for me to assess if the package contains all the functionality that is claimed or if it is working correctly.

Part of: https://github.com/openjournals/joss-reviews/issues/7435
