### 2025-01-07 c22f88d
Add helper functions to enable local testing
