+++ tests/testthat/teardown.R
@@ -1,0 +1,1 @@
+unlink(oms_path, recursive = TRUE)
