### 2025-01-07 ef6ff9a
Disable testing for multi OS workflow
