### 2025-01-07 2a666a4
Add testing for Mac OS only
