### 2025-01-07 2bacd9f
Update tests to work with local install
