+++ tests/testthat/test-ClassModel.R
@@ -1,9 +1,9 @@
 test_that("OpenMppModel class works", {
   skip_on_cran()
-  skip_on_os('windows')
 
   # Setup
-  use_OpenMpp_remote()
+  local_initiate_oms(oms_path)
+  use_OpenMpp_local()
   model_name <- get_models()$Name[[1]]
   model <- load_model(model_name)
 
@@ -22,5 +22,4 @@ test_that("OpenMppModel class works", {
   expect_type(model$ModelDigest, 'character')
   expect_type(model$OpenMppType, 'character')
   expect_equal(model$OpenMppType, 'Model')
-  expect_output(model$print())
 })
