### 2025-01-07 1618078
Change workflow name and fix args
