### 2025-01-07 3d63381
Updated README to include installation instructions for OpenM++
