### 2025-01-07 951fb0b
Add contribution guidelines doc
