+++ README.Rmd
@@ -22,7 +22,7 @@ The goal of `openmpp` is to provide a programmatic interface to the OpenM++ API 
 
 ## Installation (R Package)
 
-You can install the development version of `openmpp` from [GitHub](https://github.com/) with:
+You can install the development version of `openmpp` from [GitHub](https://github.com/mattwarkentin/openmpp) with:
 
 ``` r
 # install.packages("remotes")
@@ -31,40 +31,43 @@ remotes::install_github("mattwarkentin/openmpp")
 
 ## Installation (OpenM++)
 
-If you do not have access to an existing OpenM++ server, it is possible to run a server on your workstation, which the `openmpp` package can connect to.
+If you do not have access to an existing OpenM++ server, you can download and install OpenM++ locally and run a local server on your workstation. The `openmpp` package can then connect to this local server instance.
 
 ### Download
 
-To install OpenM++, download and unzip the "source code and binaries" appropriate for your system from [GitHub](https://github.com/openmpp/main/releases/latest)
+For most users, the best way to install OpenM++ locally is to download the pre-compiled binaries. To install OpenM++, download and unzip the "Source code and binaries" appropriate for your operating system. The latest release of OpenM++ can be found here: <https://github.com/openmpp/main/releases/latest>. Pre-compiled binaries are available for Mac (Intel and Arm), Windows, and several common Linux distributions (Debian, Ubuntu, RedHat).
 
-NOTE: Windows may allow you to view the contents of the zip file without extracting, however the files must be extracted for proper function.
+NOTE: Windows may allow you to view the contents of the zip directory without extracting, however, the files must be extracted for the installation to function properly.
 
-### Running on Windows
+### Running OpenM++ on Windows
 
-Enter the OpenM++ directory using File Explorer. Right-click inside the folder and select "Open in Terminal".
+Enter the OpenM++ directory using the File Explorer. Right-click anywhere inside the folder and select "Open in Terminal".
 
-In the new Windows Terminal window, type in the following command:
+In the Windows Terminal, enter the following command:
+
 ``` cmd
 .\bin\oms.exe
 ```
-This will start the process responsible for running the web service and running models. Take note of the address listed after "Listen at:", as this will be the `OPENMPP_LOCAL_URL` to be used with the R package.
 
-### Running on MacOS
+This will start the process responsible for running the web service (OMS). Note that the address listed after __"Listen at:"__ is the local host address that will be used by the `openmpp` R package to communicate with the API. This address will be used as the `OPENMPP_LOCAL_URL`. See the Usage section for more details.
 
-Open a new MacOS Terminal window (either by using Spotlight search or by navigating to "Applications" and then "Utilities" in Finder).
+### Running OpenM++ on MacOS
 
-After unzipping the downloaded file in Finder, drag the folder into the terminal and press Enter. This will change your active directory to the OpenM++ folder.
+Open a new MacOS Terminal window (either by using Spotlight Search or by navigating to "Applications" and then "Utilities" in Finder).
 
-Type the following command into the terminal:
+After unzipping the downloaded directory in Finder, drag the folder into the terminal and press Enter. This will change your active directory to the OpenM++ folder.
+
+Enter the following command into the terminal:
+
 ``` bash
 bin/oms
 ```
-Like the Windows installation, the web service will begin running and present the "Listen at:" address for use in the R package.
 
+Similar to the Windows installation, the web service (OMS) will initiate and the local host address will be shared with the R package for API communication.
 
 ## Usage
 
[…]
