### 2025-01-07 8d5f742
Add and remove some API endpoints
