+++ R/APIConnections.R
@@ -1,14 +1,15 @@
 #' OpenM++ API Connections
 #'
-#' Register a connection to the OpenM++ Application Programming Interface (API).
-#'   Currently, two different API connections "local" and "remote" are
-#'   available. Note that "local" and "remote" describe where the API is running
-#'   relative to the machine running the R session. For example, users logged
-#'   into the cloud-hosted RStudio Server sessions will acceess the API running
-#'   locally. Those who use a machine running on their local network may use
-#'   a remote connection to connect with the cloud-based API.
+#' Register a connection to the OpenM++ Web Services (OMS) Application
+#'   Programming Interface (API). Currently, two different API connections
+#'   "local" and "remote" are available. Note that "local" and "remote"
+#'   describe where the API is running relative to the machine running the R
+#'   session. Those who use a machine running on their local network may use
+#'   a remote connection to connect with a cloud-based API. Users running
+#'   OpenM++ locally or who are logged into a remote virtual machine running
+#'   OpenM++ will use the local API connection.
 #'
-#' @param url URl for making API requests. See `Details` for more
+#' @param url URL for making API requests. See `Details` for more
 #'   instructions.
 #' @param user User name for logging into remote API. See `Details` for more
 #'   instructions.
