### 2025-01-07 mattwarkentin/openmpp#15 @mattwarkentin
Thanks for the comment. We have now added a [CONTRIBUTING.md](https://github.com/mattwarkentin/openmpp/blob/main/CONTRIBUTING.md) document which details how interested parties can contribute to this package or seek help. We have also included the following statement in the README to point users to this document:
https://github.com/mattwarkentin/openmpp/blob/8d5f742dd9296f932557ae1c4955f05464cbab42/README.Rmd#L297-L299
