### 2025-01-07 mattwarkentin/openmpp#11 @mattwarkentin
> If the intended audience of this package includes users who are new to OpenM++, then the [installation instructions](https://github.com/mattwarkentin/openmpp?tab=readme-ov-file#user-authentication) could be improved. In particular, more guidance directly to download and start up instructions for OpenM++ would be helpful...

Thanks for this comment. We agree that the Installation instructions were insufficient. We have several sections to the README file to provide greater details on how users can download and install OpenM++ to run it locally and use this package more easily:
https://github.com/mattwarkentin/openmpp/blob/cde0b96b6675d5a04ad5e7b40c91f80deab70c52/README.md?plain=1#L15-L81

> more direct access to the .Renviron file

We have added the following statement to the README to help guide users in setting environment variables in their `.Renviron` file using the established `{usethis}` R package.
https://github.com/mattwarkentin/openmpp/blob/cde0b96b6675d5a04ad5e7b40c91f80deab70c52/README.md?plain=1#L96-L99

> examples of where to find the local API address

We provide descriptions of how to find the local host address in the installation instructions and the usage sections:

https://github.com/mattwarkentin/openmpp/blob/cde0b96b6675d5a04ad5e7b40c91f80deab70c52/README.md?plain=1#L57-L62

https://github.com/mattwarkentin/openmpp/blob/cde0b96b6675d5a04ad5e7b40c91f80deab70c52/README.md?plain=1#L79-L81

https://github.com/mattwarkentin/openmpp/blob/cde0b96b6675d5a04ad5e7b40c91f80deab70c52/README.md?plain=1#L106-L108
