### 2025-01-07 mattwarkentin/openmpp#12 @mattwarkentin
We now provide some more context and a link for even further details about the model demonstrated in the Examples section:

https://github.com/mattwarkentin/openmpp/blob/cde0b96b6675d5a04ad5e7b40c91f80deab70c52/README.md?plain=1#L190-L202
