### 2025-01-07 mattwarkentin/openmpp#17 @mattwarkentin
> However, it's not clear that this example provides an exhaustive demonstration of the package functionality.

The Example model (RiskPaths) show in the documentation is a fairly simple microsimulation model that is included/shipped with the software precisely because it is small and not very complex. While production-scale models will have much greater complexity, we think the general principles of how to use the package that are shown in the Example should get people on the right track for working with THEIR models.

In real-world use of this package, we use this package to interact with models that have hundreds of input parameter tables, very complicated simulation pathways, and hundreds of output tables. However, we use the same general workflow principles demonstrated in the example. It is difficult to demonstrate more complex models without getting lost in the minutia of models that won't be available or relevant to other users.
