### 2025-01-07 mattwarkentin/openmpp#16 @mattwarkentin
Thank you for the comment, we have added the following statement to the [paper.md]:

> To our knowledge, there are no other R packages that provide programmatic access to an open-source software for dynamic microsimulation.
