+++ R/oms-ModelingTaskMetadata.R
@@ -1,9 +1,9 @@
-#' OpenM++ Model Task
+#' Modeling Task Metadata
 #'
 #' Functions for getting creating, updating, retrieving, and deleting
-#'   model tasks.
+#'   model tasks. More information about these API endpoints can be found at
+#'   [here](https://github.com/openmpp/openmpp.github.io/wiki/Oms-web-service-API#get-modeling-task-metadata-and-task-run-history).
 #'
-#' @param data Data used for the body of the request.
 #' @inheritParams get_model
 #' @inheritParams get_workset_param
 #' @inheritParams get_run_microdata
@@ -106,39 +106,4 @@ get_model_task_run_compl <- function(model, task) {
     httr2::req_url_path(api_path) |>
     httr2::req_perform() |>
     httr2::resp_body_json()
-}
-
-#' @rdname get_model_task
-#' @export
-create_task <- function(data) {
-  api_path <- glue::glue('/api/task-new')
-  OpenMpp$API$build_request() |>
-    httr2::req_url_path(api_path) |>
-    httr2::req_body_json(data, auto_unbox = TRUE) |>
-    httr2::req_method('PUT') |>
-    httr2::req_perform() |>
-    httr2::resp_body_json()
-}
-
-#' @rdname get_model_task
-#' @export
-update_task <- function(data) {
-  api_path <- glue::glue('/api/task')
-  OpenMpp$API$build_request() |>
-    httr2::req_url_path(api_path) |>
-    httr2::req_body_json(data, auto_unbox = TRUE) |>
-    httr2::req_method('PATCH') |>
-    httr2::req_perform() |>
-    httr2::resp_body_json()
-}
-
-#' @rdname get_model_task
-#' @export
-delete_task <- function(model, task) {
-  api_path <- glue::glue(' /api/model/{model}/task/{task}')
-  OpenMpp$API$build_request() |>
-    httr2::req_url_path(api_path) |>
-    httr2::req_method('DELETE') |>
-    httr2::req_perform()
-  invisible()
 }
