### 2025-01-08 mattwarkentin/openmpp#14 @mattwarkentin
> Documentation of openmpp functions could be improved. Currently it is quite terse.

Many/most of the functions exported by this package are very low-level API endpoints that the vast majority of users won't/shouldn't use directly. But we do export them for completeness. Those who seek to use these functions will likely be more expert.

The main classes (`OpenMppModel`, `OpenMppWorkset`, `OpenMppModelRun`, `OpenMppModelRunSet`) and commonly-used functions (e.g., `load_model()`, etc.) are documented fairly extensively.

> Many documentation files also include a long list of usage examples, but do not describe what these different uses do. e.g.

These files which contain many usage examples are the low-level functions described above that we don't really expect most users will ever need to use directly. We believe that those users that seek to use these functions directly will have an expert understanding of when and how to use them.

However, to improve the ability for users to understand these low-level functions, we have updated all of our documentation to include links to the source documentation for these API endpoints. For example:

https://github.com/mattwarkentin/openmpp/blob/8774311d5bd1600101ad5dbee090a994450c7642/R/oms-Administrative.R#L1-L4
