+++ man/openmpp-package.Rd
@@ -6,7 +6,7 @@
 \alias{openmpp-package}
 \title{openmpp: Programmatic Interface to OpenM++}
 \description{
-A programmatic interface to the OpenM++ microsimulation platform.
+A programmatic interface to the OpenM++ microsimulation platform. The primary goal of this package is to wrap the OpenM++ Web Service (OMS) to provide existing OpenM++ users a programmatic interface for the R language.
 }
 \seealso{
 Useful links:
