### 2025-01-08 mattwarkentin/openmpp#13 @mattwarkentin
Thank you for the comment, we agree that this clarification is important. We have added the following sentence to the `DESCRIPTION` file for the R package and to the JOSS manuscript.

> The primary goal of this package is to wrap the OpenM++ Web Service (OMS) to
    provide existing OpenM++ users a programmatic interface for the R language.

For example:

https://github.com/mattwarkentin/openmpp/blob/abc7a0f09a6e22bf2eb01394a1766149235bc32a/DESCRIPTION#L7-L9
