### 2025-01-08 mattwarkentin/openmpp#9 @mattwarkentin
Thanks for the detailed comments, @agerada. Please see my responses:

> 1. Nomenclature.

You are correct that there is a package named `openMpp`. This is a legacy R package which is largely abandoned (as far as I can tell). It was developed by the author/maintainer of OpenM++ but really falls short of providing complete and easy-to-use functionality for the user base. My R package (`openmpp`) is now recommend on their website and under active development and maintenance. My package took an opinionated approach of using `R6` for the programmatic interface and it also provides much more complete coverage of API endpoints, user friendly functions, documentation, continuous integration and testing so I think it distinctly stands apart from this prior work.

> 2. Installation and getting started.

We completely agree that the installation instructions were insufficient for new users of this package to get started. Please see #11 for all the ways in which we improved the documentation to detail installation instructions.

> 3. Use of R6 as an OOP system.

This is a tough one. I agree that choosing this approach is not very idiomatic of R but I did/do believe it offers a nice approach for an API-wrapping package like this one. Since this package relies on communicating with external software/databases/API, the OOP approach of R6 allows for interacting with these using well-defined (and user friendly) methods/functions. We do provide functions that allow users to use the lower-level API calls, so we haven't necessarily limited the use of more idiomatic R code, but we do like the workflow allowed by the use of R6 OOP.

We do go to length in the manuscript why we chose the R6 approach (maybe too much?). I have now added to the README to expand on the choice of using R6.

https://github.com/mattwarkentin/openmpp/blob/34a3b2b2785a8cc9ca0f8f0e154496e1ef9c1057/README.md?plain=1#L184-L196

> 4. Testing suite. Could @mattwarkentin clarify, is that using a web-hosted OpenM++ instance or a temporary one?

I have now fixed the testing to use a fresh local installation each time testing is ran. It no longer relies on remote access to an OpenM++ that no one has access to. For now, it relies on installing locally a MacOS binary since this is my development environment and can be used nicely with GitHub Actions.

I have now added a new GitHub Action for testing the package here on a MacOS runner here: https://github.com/mattwarkentin/openmpp/blob/main/.github/workflows/tests.yaml

The package is still checked for installation using R CMD CHECK on MacOS, Linux, and Windows using the original GitHub Action here: https://github.com/mattwarkentin/openmpp/blob/main/.github/workflows/R-CMD-check.yaml

> 5. Developer notes. I cannot find a set of developer notes on how to set up a local developer instance and test the package. That would be useful for users intending to contribute.

Following from the above, testing is more robust now. Developers _should_ be able to clone the repository and test it using the standard `{testthat}` testing infrastructure for R packages. However, right now it will only work for someone using Arm MacOS since I've configured it to work in my dev environment (and a MacOS GitHub Action). I think for it to work for "anyone", I would have to added a good deal more code in [`tests/testthat/setup.R`](https://github.com/mattwarkentin/openmpp/blob/main/tests/testthat/setup.R) to detect the host OS and download/install/run the right pre-compiled binary. Its no impossible, but its not nothing either.

It might be worth doing this so that the package testing can also be done across all OS's. I'm very open to suggestions on this one.

> The link to github [here](https://github.com/mattwarkentin/openmpp/?tab=readme-ov-file#models-scenarios-runs-and-runsets) links to github's main page rather than the repo itself.

Good catch. Fixed.

> Can you please confirm whether there is a guideline for contributions to this software as per JOSS's requirements

Thanks for the comment. We have now added a [CONTRIBUTING.md](https://github.com/mattwarkentin/openmpp/blob/main/CONTRIBUTING.md) document which details how interested parties can contribute to this package or seek help. We have also included the following statement in the README to point users to this document:
https://github.com/mattwarkentin/openmpp/blob/8d5f742dd9296f932557ae1c4955f05464cbab42/README.Rmd#L297-L299
