+++ inst/paper/paper.md
@@ -34,10 +34,10 @@ Dynamic microsimulation is a computational technique used to simulate the behavi
 
 [OpenM++](https://openmpp.org/index.html) is an open-source software platform designed for dynamic microsimulation modeling [@OpenMpp]. It provides a flexible framework and language for researchers and policymakers to build, calibrate, and analyze complex dynamic microsimulation models. OpenM++ offers a user-friendly interface and a powerful domain-specific language, allowing users to define intricate demographic, economic, and behavioral processes within a simulated population. With its modular architecture, OpenM++ enables easy integration of various data sources and model components, facilitating interdisciplinary research and policy analysis. The OpenM++ platform is widely used in academia, government agencies, and research institutions to explore the long-term impacts of policy interventions, demographic changes, and economic trends on populations and societies. However, while the OpenM++ graphical interface is simple to learn, the graphical interface may impede users with more complicated modelling needs and limits the transparency and reproducibility of analyses. The OpenM++ software provides an application programming interface (API) that enables programmatic interactions with data and models using scripting languages (e.g., R).
 
-The [`openmpp`](https://github.com/mattwarkentin/openmpp) R package provides a programmatic interface to the OpenM++ API [@openmpp] to simplify the process of creating and managing microsimulations, performing scalable and reproducible computations, and gathering results adhereing to tidy data principles [@tidyverse] for further processing and reporting. OpenM++ can run on local machines, servers, or cloud infrastructure (i.e., remotely). [`openmpp`](https://github.com/mattwarkentin/openmpp) supports connectivity and secure user authentication for both local and remote connections to OpenM++ from an R session. A user may wish to define many scenarios to evaluate using dynamic microsimulation. Manual approaches to managing these scenarios and runs can be onerous and lack transparency or reproducibility. Programmatic approaches for creating and managing scenarios enables scalability and reproducibility. Boilerplate code for running microsimulations (with configurable options) encourages rapid prototyping and iteration before running simulations at scale. Lastly, output tables and reports should be kept up to date with the most recent microsimulations. The [`openmpp`](https://github.com/mattwarkentin/openmpp) package enables importing output tables directly into R to encourage reproducible reporting with no intermediate or temporary files. The package provides a suite of functions for easy communication with the OpenM++ API, including a set of R6 classes that enable real-time communication to ensure that all changes made to microsimulation objects in the R session are propagated and synchronized with the OpenM++ database. R6 provides the infrastructure to abstract away the technical debt required to send and receive data when communicating with the API [@R6]. Users can interact with R objects in a familiar way while achieving real-time synchronization with external resources. Encapsulated object-oriented programming (OOP) allows the internal state of the object (i.e., the connection to the actual object in the OpenM++ database) to be accessed and modified only through well-defined methods, rather than directly manipulating the data, which enforces data integrity, improves code readability, and simplifies maintenance by abstracting away the implementation details of an object and preventing unintended modifications to its state.
[…]
