### 2025-01-08 mattwarkentin/openmpp#10 @mattwarkentin
> 1. The final section on R6 OOP (lines 61-69) feels a little out of place. I'm not sure it adds much to the paper, but if you do keep it in, I suggest adding a few lines of example code (e.g., load model, load scenario, and execute run).

I wasn't sure how to handle the discussion around `R6`. The design principle of the package relies heavily on `R6` and (as you pointed out) this is not really idiomatic R code. So, I felt some need to justify WHY I chose to use `R6`. But perhaps I could trim this section down. In any case, I have added an abbreviated version of this discussion to the README so users will be confronted by this decision-making regardless off if they read the manuscript, which I think is valuable based on your suggestion in #9.

I was hesitant to include any code in the manuscript, since this is something that will evolve during the life of the package. I borrowed inspiration for this manuscript from https://github.com/ropensci/targets/blob/main/inst/paper/paper.md. Basically just introduce the package, why it was designed the way it was, and why someone would use it, without committing to showing code since this liable to change. Example code are shared in the GitHub repository and package website since this is something I can update over time, whereas the manuscript will be static one published.

> 2. I think more could be added to place this software in the context of related work. Although the paper does say that users can use OpenM++'s GUI as an alternative, I would suggest mentioning other alternative scripting approaches to programmatically control modelling, e.g., other R or python packages that interact with OpenM++.

Thanks for this comment. In principle, any language that allows communication with an API (all programming languages?) are permissible for interacting with OpenM++. However, this will require signifiant work on the part of the user. This is why I developed this package for R users, but similar approaches could be adopted by users of Python or other languages, but this is beyond the scope of this project. Again, open to suggestions about how best to address this in the documentation or manuscript.

The comment about placing this in the context of related work is important. We do not know of any other work/packages similar to this. So we have added a statement reflecting this to the manuscript:

> The primary goal of this package is to wrap the OpenM++ Web Service (OMS) to provide existing OpenM++ users a programmatic interface for the R language. To our knowledge, there are no other R packages that provide programmatic access to an open-source software for dynamic microsimulation.
