### 2025-01-08 #7435 @mattwarkentin
Hi @mikemahoney218, @evaenns, @agerada,

Thanks for the detailed package and manuscript reviews. I have tried to address each of the comments within the corresponding issues over in the [mattwarkentin/openmpp](https://github.com/mattwarkentin/openmpp/issues) repository.

Please let me know how we proceed/what happens next. Thanks again for your feedback. I think it has helped to improve the package and manuscript.
