### 2025-01-08 b451e86
Update JOSS manuscript
