+++ .github/workflows/R-CMD-check.yaml
@@ -44,7 +44,12 @@ jobs:
           extra-packages: any::rcmdcheck
           needs: check
 
+      - name: Setup .Renviron for testing
+        run: |
+          echo OPENMPP_LOCAL_URL=http://localhost:4040 >> ~/.Renviron
+        shell: bash
+
       - uses: r-lib/actions/check-r-package@v2
         with:
-          args: 'c("--no-manual", "--as-cran", "--no-tests")'
+          args: 'c("--no-manual", "--as-cran")'
           upload-snapshots: true
