### 2025-01-08 b2237e2
Remove Mac only testing action
