### 2025-01-08 9e1ad4b
Fix linux openmpp installer
