### 2025-01-08 e85b375
Fixed Windows testing
