### 2025-01-08 ae529dd
Default to using 4040 port for local API connection
