+++ man/use_OpenMpp_local.Rd
@@ -52,7 +52,8 @@ sensitive information, keeping all of this information in one place makes
 sense for consistency. For local and remote API connections, set the
 following environment variables in your \code{.Renviron} files:
 \itemize{
-\item \code{OPENMPP_LOCAL_URL}: URL for a local API connection.
+\item \code{OPENMPP_LOCAL_URL}: URL for a local API connection. Default is to use
+http://localhost:4040.
 \item \code{OPENMPP_REMOTE_URL}: URL for a remote API connection.
 \item \code{OPENMPP_REMOTE_USER}: User name for logging into a remote API
 connection.
