### 2025-01-08 716b149
Simplify if statements in test setup
