### 2025-01-09 9e02295
Safer way to always get binaries from latest release
