+++ tests/testthat/setup.R
@@ -1,19 +1,10 @@
 local_install_openmpp <- function() {
+  latest <- gh::gh("https://api.github.com/repos/openmpp/main/releases/latest")
   info <- Sys.info()
   os <- info['sysname']
   type <- info['machine']
 
-  if (os == 'Darwin' & type == 'arm64') {
-    url <- 'https://github.com/openmpp/main/releases/latest/download/openmpp_mac_arm64_20241226.tar.gz'
-  } else if (os == 'Darwin' & type == 'x86_64') {
-    url <- 'https://github.com/openmpp/main/releases/latest/download/openmpp_mac_x86_64_20241226.tar.gz'
-  } else if (os == 'Linux') {
-    url <- 'https://github.com/openmpp/main/releases/latest/download/openmpp_debian-11_20241226.tar.gz'
-  } else if (os == 'Windows') {
-    url <- 'https://github.com/openmpp/main/releases/latest/download/openmpp_win_20241226.zip'
-  } else {
-    rlang::abort('No compatible OS detected.')
-  }
+  url <- find_latest_binary(latest$assets, os, type)
 
   file <- basename(url)
 
@@ -33,6 +24,28 @@ local_install_openmpp <- function() {
   }
 
   new_path
+}
+
+find_latest_binary <- function(assets, os, type) {
+  asset_names <- purrr::map_chr(assets, \(x) x$name)
+
+  if (os == 'Darwin' & type == 'arm64') {
+    idx <- which(grepl('openmpp_mac_arm64_\\d{8}.tar.gz', asset_names))
+    url <- assets[[idx]]$browser_download_url
+  } else if (os == 'Darwin' & type == 'x86_64') {
+    idx <- which(grepl('openmpp_mac_x86_64_\\d{8}.tar.gz', asset_names))
+    url <- assets[[idx]]$browser_download_url
+  } else if (os == 'Linux') {
+    idx <- which(grepl('openmpp_debian-11_\\d{8}.tar.gz', asset_names))
+    url <- assets[[idx]]$browser_download_url
+  } else if (os == 'Windows') {
+    idx <- which(grepl('openmpp_win_\\d{8}.zip', asset_names))
+    url <- assets[[idx]]$browser_download_url
+  } else {
+    rlang::abort('No compatible OS detected.')
+  }
+
+  url
 }
 
 oms_path <- local_install_openmpp()
