### 2025-01-09 15ce486
Better way to check port is listening before running tests
