### 2025-02-13 0da9f92
Rewrite to the API connections interface to generalize future new connections
