+++ NAMESPACE
@@ -133,6 +133,7 @@ export(update_workset_param)
 export(update_workset_param_csv)
 export(upload_user_files)
 export(upload_workset_params)
+export(use_OpenMpp_CPAC)
 export(use_OpenMpp_local)
 export(use_OpenMpp_remote)
 import(cli)

[…]
