+++ R/API-CPAC.R
@@ -21,14 +21,14 @@
 #'   sense for consistency. For remote API connections to CPAC, set the
 #'   following environment variables in your `.Renviron` files:
 #'
-#'   - `OPENMPP_CPAC_URL`: URL for a local API connection.
-#'   - `OPENMPP_CPAC_API`: URL for a remote API connection.
+#'   - `OPENMPP_CPAC_URL`: URL for logging into a remote API connection.
+#'   - `OPENMPP_CPAC_API`: URL for making remote API requests.
 #'   - `OPENMPP_CPAC_USER`: User name for logging into the remote API
 #'     connection.
 #'   - `OPENMPP_CPAC_PWD`: Password for logging into the remote API connection.
 #'
 #' @return Nothing, invisibly. Behind-the-scenes, an instance of the
-#'   `OpenMppCPAC` R6 classe is created. This objects should not be accessed
+#'   `OpenMppCPAC` R6 class is created. This objects should not be accessed
 #'   directly by the user, instead, the package internally uses these
 #'   connections to communicate with the OpenM++ API.
 #'
