### 2025-03-25 mattwarkentin/openmpp#9 @agerada
Thank you @mattwarkentin and again apologies for the delay in replying to your comments.

I have no further comments as you have addressed all my previous points.

I did have an issue with testing again, unfortunately `openmpp` binaries for Mac OS were not available for the latest release. I do not know how often this happens, but I have submitted a pull request #19 to hopefully make the automated download of binaries (which I like) more robust.

My only final (optional) question is whether you are planning to submit to `CRAN`? While not completely necessary I think it will help access to the package. The local `R CMD check` seems to run without any errors/warnings so I don't think there would be any issues getting this on cran.
