### 2025-03-25 mattwarkentin/openmpp#10 @agerada
Thank you @mattwarkentin. I agree with your points, in particular the avoidance of code in the paper (which I believe is in line with JOSS's suggestions).

One final (optional) point is whether you have examples/references of the package being used in specific research or other outputs. If none yet, then ignore.
