### 2025-03-25 #7435 @agerada
Hi @crvernon, I have had some time today to complete this review. I am happy with the changes and comments provided by @mattwarkentin regarding both the software and the paper. Therefore, I have ticked off the remaining items on my checklist and can consider the review complete. There are some minor questions/PRs that I have left with @mattwarkentin but they do not change the outcome of the review.

Thank you for the invitation to review.
