### 2025-03-31 mattwarkentin/openmpp#10 @mattwarkentin
Great suggestion, I think we have some published articles that I could cite as examples of using this package.
