### 2025-03-31 mattwarkentin/openmpp#10 @mattwarkentin
Note to self:
```
@article{basmadjian2025examining,
  title={Examining breast cancer screening recommendations in Canada: The projected resource impact of screening among women aged 40--49},
  author={Basmadjian, Robert B and Ruan, Yibing and Hutchinson, John M and Warkentin, Matthew T and Alagoz, Oguzhan and Coldman, Andrew and Brenner, Darren R},
  journal={Journal of Medical Screening},
  volume={32},
  number={1},
  pages={35--43},
  year={2025},
  publisher={SAGE Publications Sage UK: London, England}
}
```
