### 2025-03-31 mattwarkentin/openmpp#10 @mattwarkentin
```
Manuscript Number: EPI-24-1488R3
Manuscript Type: Research Article
Manuscript Title: Modeling the Economic and Health Impact of Lowering the Recommended Colorectal Cancer Screening Age in Canada using Fecal Immunochemical Test versus Colonoscopy
Corresponding Author: Dr. Loree
Full Author List: Brendan J. Chia, Yibing Ruan, Carl J Brown, Robert J Hilsden, John M Hutchinson, Darren R Brenner, and Jonathan M. Loree
```
