### 2025-04-01 mattwarkentin/openmpp#9 @mattwarkentin
Thanks for the contribution! Merged #19
