### 2025-04-01 2861255
Cite published studies that have used openmpp for microsim
