+++ inst/paper/paper.md
@@ -34,7 +34,7 @@ Dynamic microsimulation is a computational technique used to simulate the behavi
 
 [OpenM++](https://openmpp.org/index.html) is an open-source software platform designed for dynamic microsimulation modeling [@OpenMpp]. It provides a flexible framework and language for researchers and policymakers to build, calibrate, and analyze complex dynamic microsimulation models. OpenM++ offers a user-friendly interface and a powerful domain-specific language, allowing users to define intricate demographic, economic, and behavioral processes within a simulated population. With its modular architecture, OpenM++ enables easy integration of various data sources and model components, facilitating interdisciplinary research and policy analysis. The OpenM++ platform is widely used in academia, government agencies, and research institutions to explore the long-term impacts of policy interventions, demographic changes, and economic trends on populations and societies. However, while the OpenM++ graphical interface is simple to learn, the graphical interface may impede users with more complicated modelling needs and limits the transparency and reproducibility of analyses. The OpenM++ software provides an application programming interface (API) that enables programmatic interactions with data and models using scripting languages (e.g., R).
 
-The [`openmpp`](https://github.com/mattwarkentin/openmpp) R package provides a programmatic interface to the OpenM++ API [@openmpp] to simplify the process of creating and managing microsimulations, performing scalable and reproducible computations, and gathering results adhering to tidy data principles [@tidyverse] for further processing and reporting. The primary goal of this package is to wrap the OpenM++ Web Service (OMS) to provide existing OpenM++ users a programmatic interface for the R language. To our knowledge, there are no other R packages that provide programmatic access to an open-source software for dynamic microsimulation. 
+The [`openmpp`](https://github.com/mattwarkentin/openmpp) R package provides a programmatic interface to the OpenM++ API [@openmpp] to simplify the process of creating and managing microsimulations, performing scalable and reproducible computations, and gathering results adhering to tidy data principles [@tidyverse] for further processing and reporting. The primary goal of this package is to wrap the OpenM++ Web Service (OMS) to provide existing OpenM++ users a programmatic interface for the R language. To our knowledge, there are no other R packages that provide programmatic access to an open-source software for dynamic microsimulation. To date, several published studies have performed their microsimulations using the `openmpp` package [@basmadjian2025examining; @banik2025shooting; @chia2025modeling].
 
[…]
