+++ inst/paper/paper.bib
@@ -36,3 +36,34 @@
     pages = {1686},
     doi = {10.21105/joss.01686},
 }
+
+@article{basmadjian2025examining,
+  title={Examining breast cancer screening recommendations in Canada: The projected resource impact of screening among women aged 40--49},
+  author={Basmadjian, Robert B and Ruan, Yibing and Hutchinson, John M and Warkentin, Matthew T and Alagoz, Oguzhan and Coldman, Andrew and Brenner, Darren R},
+  journal={Journal of Medical Screening},
+  volume={32},
+  number={1},
+  pages={35--43},
+  year={2025},
+  publisher={SAGE Publications Sage UK: London, England}
+}
+
+@article{banik2025shooting,
+  title={Shooting for the Moon: Can We Cut Cancer Mortality in Canada By 50\% By 2050?},
+  author={Banik, Keaton and Ruan, Yibing and Stephen, Mariet M and Hutchinson, John M and Carbonell, Chantelle and Warkentin, Matthew T and Coldman, Andrew and Garner, Rochelle and Jalal, Hawre and Brenner, Darren R},
+  journal={Cancer Control},
+  volume={32},
+  pages={10732748251319485},
+  year={2025},
+  publisher={SAGE Publications Sage CA: Los Angeles, CA}
+}
+
+@article{chia2025modeling,
+  title={Modeling the Economic and Health Impact of Lowering the Recommended Colorectal Cancer Screening Age in Canada using Fecal Immunochemical Test versus Colonoscopy},
+  author={Chia, Brendan J and Ruan, Yibing and Brown, Carl J and Hilsden, Robert J and Hutchinson, John M and Brenner, Darren R and Loree, Jognathan M},
+  journal={Cancer Epidemiology, Biomarkers \& Prevention},
+  volume={TBD},
+  pages={TBD},
+  year={2025},
+  publisher={American Association for Cancer Research}
+}
