### 2025-04-01 mattwarkentin/openmpp#10 @mattwarkentin
Cited published studies that used `openmpp` in the JOSS manuscript in 286125587afb478f88788bd167bda07e174a262c.

Thanks again for the suggestion.
