### 2025-04-01 #7435 @mattwarkentin
I have now merged the PR to improve the testing suite (thanks to @agerada for the contribution) and updated the manuscript based on the suggestion to include some published examples of studies that used this package.

I think we are just waiting to hear back from @evaenns about whether I have adequately address their comments.
