+++ inst/paper/paper.bib
@@ -60,10 +60,8 @@
 
 @article{chia2025modeling,
   title={Modeling the Economic and Health Impact of Lowering the Recommended Colorectal Cancer Screening Age in Canada using Fecal Immunochemical Test versus Colonoscopy},
-  author={Chia, Brendan J and Ruan, Yibing and Brown, Carl J and Hilsden, Robert J and Hutchinson, John M and Brenner, Darren R and Loree, Jognathan M},
+  author={Chia, Brendan J and Ruan, Yibing and Brown, Carl J and Hilsden, Robert J and Hutchinson, John M and Brenner, Darren R and Loree, Jonathan M},
   journal={Cancer Epidemiology, Biomarkers \& Prevention},
-  volume={TBD},
-  pages={TBD},
   year={2025},
   publisher={American Association for Cancer Research}
 }
