### 2025-04-30 #7435 @crvernon
Thanks @tkerwin!  You can generate your checklist by running the following:

`@editorialbot generate my checklist`
