### 2025-05-08 #7435 @tkerwin
Due to the circumstances, it looks like all the substantial issues have already been taken care of. This seems like a useful package if you wish to interact with OpenM++ through R. I ran through the examples  (Windows 11) and everything worked as expected.

I have one minor terminology complaint: I'm not sure if the phrase "technical debt" means what you intend it to mean in the article. Technical debt refers to the extra work that must be done later in exchange for taking the more expedient route in developing a system. It seems like you just mean something like "boilerplate code" or "overly verbose code"
