### 2025-05-08 4c86724
JOSS manuscript updates
