+++ inst/paper/paper.md
@@ -26,17 +26,17 @@ bibliography: paper.bib
 
 # Summary
 
-OpenM++ is a powerful open-source and cross-platform software and domain-specific language to develop, use, and deploy dynamic microsimulations or agent-based models. The [`openmpp`](https://github.com/mattwarkentin/openmpp) R package provides a programmatic interface to OpenM++ to simplify the process of connecting to local or remote servers with secure user authentication, creating, managing, and running large-scale and complex microsimulations, performing scalable and reproducible computations, and gathering results for reproducible reporting.
+OpenM++ is a powerful open-source and cross-platform software and domain-specific language to develop, use, and deploy dynamic microsimulations or agent-based models. The [`openmpp`](https://github.com/mattwarkentin/openmpp) R package provides a programmatic interface to OpenM++ to simplify the process of connecting to local or remote instances with secure user authentication, communicating with web services, creating, managing, and running large-scale and complex microsimulations, performing scalable and reproducible computations, and gathering results for reproducible reporting.
 
 # Statement of Need
 
-Dynamic microsimulation is a computational technique used to simulate the behavior of individual entities (i.e., micro) within a population over time. It combines elements of microeconomics, demography, and statistical modeling to create detailed simulations of complex systems. Dynamic microsimulation involves creating a synthetic population representative of the real population and modeling the life paths of individuals based on various demographic and socioeconomic characteristics. Unlike static models that provide snapshots of a population at a single point in time, dynamic microsimulation models capture the interactions and changes that occur over time, allowing researchers to analyze the long-term effects of demographic shifts, policy changes, as well as clinical or population intervention. These models can then be used to predict future population trends and assess the impact of policy changes or interventions on health, resource, and economic outcomes. Dynamic microsimulation is used across a wide range of disciplines, including economics, public policy, epidemiology, urban planning, and environmental science. In public health research, dynamic microsimulation models are used to forecast disease trends, evaluate the impact of healthcare interventions, and plan for future healthcare needs. Overall, dynamic microsimulation provides a powerful framework for understanding the complex dynamics of trends within a population to inform evidence-based decision-making across a variety of fields.
+Dynamic microsimulation is a computational technique used to simulate the behavior of individual entities (i.e., micro) within a population over time. It combines elements of microeconomics, demography, and statistical modeling to create detailed simulations of complex systems. Dynamic microsimulation involves creating a synthetic population representative of a real population and modeling the life paths of individuals based on various demographic and socioeconomic characteristics. Unlike static models that provide snapshots of a population at a single point in time, dynamic microsimulation models capture the interactions and changes that occur over time, allowing researchers to analyze the long-term effects of demographic shifts, policy changes, as well as clinical or population interventions. These models can then be used to predict future population trends and assess the impact of policy changes or interventions on health, resource, and economic outcomes. Dynamic microsimulation is used across a wide range of disciplines, including economics, public policy, epidemiology, urban planning, and environmental science. In public health research, dynamic microsimulation models are used to forecast disease trends, evaluate the impact of healthcare interventions, and plan for future healthcare needs. Overall, dynamic microsimulation provides a powerful framework for understanding the complex dynamics of trends within a population to inform evidence-based decision-making across a variety of fields.
 
[…]
