+++ inst/paper/paper.md
@@ -36,7 +36,7 @@ Dynamic microsimulation is a computational technique used to simulate the behavi
 
 The [`openmpp`](https://github.com/mattwarkentin/openmpp) R package provides a programmatic interface to OpenM++ [@openmpp] to simplify the process of creating and managing microsimulations, performing scalable and reproducible computations, and gathering results adhering to tidy data principles [@tidyverse] for further processing and reporting. This is achieved by wrapping the OpenM++ Web Service (OMS) API to provide OpenM++ users a programmatic interface for the R language. To our knowledge, there are no other R packages that provide programmatic access to an open-source software for dynamic microsimulation. To date, several published studies have performed their microsimulations using the `openmpp` package [@basmadjian2025examining; @banik2025shooting; @chia2025modeling].
 
-OpenM++ can run on local machines, servers, or cloud infrastructure (i.e., remotely). [`openmpp`](https://github.com/mattwarkentin/openmpp) supports connectivity and secure user authentication for both local and remote instances of OpenM++ from an R session. A user may wish to define many scenarios to evaluate using dynamic microsimulation. Manual approaches to managing these scenarios and runs can be onerous and lack transparency or reproducibility. Programmatic approaches for creating and managing scenarios enables scalability and reproducibility. Boilerplate code for running microsimulations (with configurable options) encourages rapid prototyping and iteration before running simulations at scale. Lastly, output tables and reports should be kept up to date with the most recent microsimulations. The [`openmpp`](https://github.com/mattwarkentin/openmpp) package enables reading output tables directly into R to encourage reproducible reporting with no intermediate or temporary files. The package provides a suite of functions for easy communication with the OpenM++ API, including a set of R6 classes that enable real-time communication to ensure that all changes made to microsimulation objects in the R session are propagated and synchronized with the OpenM++ back-end database. R6 provides the infrastructure to abstract away the boilerplate code required to send and receive data when communicating with the OpenM++ API [@R6]. Users can interact with R objects in a familiar way while achieving real-time synchronization with external resources. Encapsulated object-oriented programming (OOP) allows the internal state of the object (i.e., the connection to the actual object in the OpenM++ database) to be accessed and modified only through well-defined methods, rather than directly manipulating the data, which enforces data integrity, improves code readability, and simplifies maintenance by abstracting away the implementation details and preventing unintended modifications to its state.
[…]
