### 2025-05-13 12e42b0
Fix DOIs and cleanip bib file for JOSS paper
