+++ inst/paper/paper.bib
@@ -1,40 +1,37 @@
-@misc{
-    openmpp,
-    author = {Matthew T. Warkentin},
-    title = {{openmpp: Programmatic interface to the OpenM++ microsimulation platform}},
-    year = {2024},
-    howpublished = {https://github.com/mattwarkentin/openmpp},
-    DOI = {}
+@misc{openmpp,
+  author = {Matthew T. Warkentin},
+  title = {{openmpp: Programmatic interface to the OpenM++ microsimulation platform}},
+  year = {2024},
+  howpublished = {https://github.com/mattwarkentin/openmpp},
+  doi = {}
+}
+
+@misc{OpenMpp,
+  author = {{OpenM++ Contributors}},
+  title = {{OpenM++: open source microsimulation platform}},
+  year = {2024},
+  howpublished = {https://openmpp.org/index.html},
+  doi = {}
 }
 
 @Manual{R6,
-    title = {R6: Encapsulated Classes with Reference Semantics},
-    author = {Winston Chang},
-    year = {2021},
-    note = {R package version 2.5.1},
-    url = {https://CRAN.R-project.org/package=R6},
-    doi = {10.32614/CRAN.package.R6},
+  title = {R6: Encapsulated Classes with Reference Semantics},
+  author = {Winston Chang},
+  year = {2021},
+  note = {R package version 2.5.1},
+  url = {https://CRAN.R-project.org/package=R6},
+  doi = {10.32614/CRAN.package.R6}
 }
 
-@misc{
-    OpenMpp,
-    author = {{OpenM++ Contributors}},
-    title = {{OpenM++: open source microsimulation platform}},
-    year = {2024},
-    howpublished = {https://openmpp.org/index.html},
-    DOI = {}
-}
-
-@Article{
-    tidyverse,
-    title = {Welcome to the {tidyverse}},
-    author = {Hadley Wickham and Mara Averick and Jennifer Bryan and Winston Chang and Lucy D'Agostino McGowan and Romain François and Garrett Grolemund and Alex Hayes and Lionel Henry and Jim Hester and Max Kuhn and Thomas Lin Pedersen and Evan Miller and Stephan Milton Bache and Kirill Müller and Jeroen Ooms and David Robinson and Dana Paige Seidel and Vitalie Spinu and Kohske Takahashi and Davis Vaughan and Claus Wilke and Kara Woo and Hiroaki Yutani},
-    year = {2019},
-    journal = {Journal of Open Source Software},
-    volume = {4},
-    number = {43},
-    pages = {1686},
-    doi = {10.21105/joss.01686},
+@Article{tidyverse,
+  title = {Welcome to the {tidyverse}},
+  author = {Hadley Wickham and Mara Averick and Jennifer Bryan and Winston Chang and Lucy D'Agostino McGowan and Romain François and Garrett Grolemund and Alex Hayes and Lionel Henry and Jim Hester and Max Kuhn and Thomas Lin Pedersen and Evan Miller and Stephan Milton Bache and Kirill Müller and Jeroen Ooms and David Robinson and Dana Paige Seidel and Vitalie Spinu and Kohske Takahashi and Davis Vaughan and Claus Wilke and Kara Woo and Hiroaki Yutani},
+  year = {2019},
+  journal = {Journal of Open Source Software},
+  volume = {4},
+  number = {43},
+  pages = {1686},
+  doi = {10.21105/joss.01686}
 }
 
 @article{basmadjian2025examining,
@@ -45,7 +42,8 @@
   number={1},
   pages={35--43},
   year={2025},
-  publisher={SAGE Publications Sage UK: London, England}
[…]
