### 2025-05-13 #7435 @mattwarkentin
@crvernon Sounds good, thanks! Fixed the missing DOIs with the latest commit.
