### 2025-05-19 #7435 @crvernon
👋  @mattwarkentin - we are almost there!

Please fix the following in your paper via your bib file:

- [ ] LINE 82:  The "c" in "canada" should be capitalized.  You can do this in your bib file by placing the characters you wish to maintain formatting of in curly brackets.
- [ ] LINE 85:  The "c" in "canada" should be capitalized.
- [ ] LINE 92: The "c" in "canada" should be capitalized.

So here is what we have left to do:

- [ ]  Conduct a GitHub release of the current reviewed version of the software and archive the reviewed software in Zenodo or a similar service (e.g., figshare, an institutional repository).  Please ensure that the software archive uses the same license as the license you have posted on code repository.

- [ ]  Check the archival deposit (e.g., in Zenodo) to ensure it has the correct metadata. This includes the title (should match the paper title exactly) and author list (should match the people and the order in which they are listed on your paper exactly).

- [ ]  Please respond with the DOI of the archived version here along with the version number released.
