### 2025-05-20 04d427d
Update paper.bib to ensure proper capitalization for JOSS paper references
