+++ inst/paper/paper.bib
@@ -15,7 +15,7 @@
 }
 
 @Manual{R6,
-  title = {R6: Encapsulated Classes with Reference Semantics},
+  title = {{R6: Encapsulated Classes with Reference Semantics}},
   author = {Winston Chang},
   year = {2021},
   note = {R package version 2.5.1},
@@ -24,7 +24,7 @@
 }
 
 @Article{tidyverse,
-  title = {Welcome to the {tidyverse}},
+  title = {{Welcome to the {tidyverse}},
   author = {Hadley Wickham and Mara Averick and Jennifer Bryan and Winston Chang and Lucy D'Agostino McGowan and Romain François and Garrett Grolemund and Alex Hayes and Lionel Henry and Jim Hester and Max Kuhn and Thomas Lin Pedersen and Evan Miller and Stephan Milton Bache and Kirill Müller and Jeroen Ooms and David Robinson and Dana Paige Seidel and Vitalie Spinu and Kohske Takahashi and Davis Vaughan and Claus Wilke and Kara Woo and Hiroaki Yutani},
   year = {2019},
   journal = {Journal of Open Source Software},
@@ -35,7 +35,7 @@
 }
 
 @article{basmadjian2025examining,
-  title={Examining breast cancer screening recommendations in Canada: The projected resource impact of screening among women aged 40--49},
+  title={{Examining breast cancer screening recommendations in Canada: The projected resource impact of screening among women aged 40--49}},
   author={Basmadjian, Robert B and Ruan, Yibing and Hutchinson, John M and Warkentin, Matthew T and Alagoz, Oguzhan and Coldman, Andrew and Brenner, Darren R},
   journal={Journal of Medical Screening},
   volume={32},
@@ -47,7 +47,7 @@
 }
 
 @article{banik2025shooting,
-  title={Shooting for the Moon: Can We Cut Cancer Mortality in Canada By 50\% By 2050?},
+  title={{Shooting for the Moon: Can We Cut Cancer Mortality in Canada By 50\% By 2050?}},
   author={Banik, Keaton and Ruan, Yibing and Stephen, Mariet M and Hutchinson, John M and Carbonell, Chantelle and Warkentin, Matthew T and Coldman, Andrew and Garner, Rochelle and Jalal, Hawre and Brenner, Darren R},
   journal={Cancer Control},
   volume={32},
@@ -58,7 +58,7 @@
 }
 
 @article{chia2025modeling,
-  title={Modeling the Economic and Health Impact of Lowering the Recommended Colorectal Cancer Screening Age in Canada using Fecal Immunochemical Test versus Colonoscopy},
+  title={{Modeling the Economic and Health Impact of Lowering the Recommended Colorectal Cancer Screening Age in Canada using Fecal Immunochemical Test versus Colonoscopy}},
   author={Chia, Brendan J and Ruan, Yibing and Brown, Carl J and Hilsden, Robert J and Hutchinson, John M and Brenner, Darren R and Loree, Jonathan M},
   journal={Cancer Epidemiology, Biomarkers \& Prevention},
   year={2025},
