### 2025-05-20 4961da0
Fixed closing bracket in paper.bib
