+++ inst/paper/paper.bib
@@ -24,7 +24,7 @@
 }
 
 @Article{tidyverse,
-  title = {{Welcome to the {tidyverse}},
+  title = {{Welcome to the {tidyverse}}},
   author = {Hadley Wickham and Mara Averick and Jennifer Bryan and Winston Chang and Lucy D'Agostino McGowan and Romain François and Garrett Grolemund and Alex Hayes and Lionel Henry and Jim Hester and Max Kuhn and Thomas Lin Pedersen and Evan Miller and Stephan Milton Bache and Kirill Müller and Jeroen Ooms and David Robinson and Dana Paige Seidel and Vitalie Spinu and Kohske Takahashi and Davis Vaughan and Claus Wilke and Kara Woo and Hiroaki Yutani},
   year = {2019},
   journal = {Journal of Open Source Software},
