### 2025-05-20 #7435 @johnhutchinson99
@crvernon Thank you very much for your comments.
We've just fixed the bib file to fix the capitalization. This latest bib file is reflected in this commit: [Commit 4961da0](https://github.com/mattwarkentin/openmpp/commit/4961da0b5bb153c2c0abb439fef6e766d2722bb7)

The above PDF also reflects this change. We will handle each of the other points soon and will notify you when it is finished.
