### 2025-05-21 #7435 @mattwarkentin
Hi @crvernon,

- [X] Conduct a GitHub release of the current reviewed version of the software and archive the reviewed software in Zenodo or a similar service (e.g., figshare, an institutional repository). Please ensure that the software archive uses the same license as the license you have posted on code repository.
- [X] Check the archival deposit (e.g., in Zenodo) to ensure it has the correct metadata. This includes the title (should match the paper title exactly) and author list (should match the people and the order in which they are listed on your paper exactly).
- [X] Please respond with the DOI of the archived version here along with the version number released.

DOI: https://doi.org/10.5281/zenodo.15484568
Zenodo: https://zenodo.org/records/15484568
Software version: 0.0.1

I think this covers everything now. Please let us know if you need anything else from our end.
